nutdrv_qx_masterguard: Adds battery test until battery low.
The difference between the battery test until battery low and
deep battery test is that the latter is actually a battery calibration
test and needs some preconditions to be met, i.e. the load must be
between 30% and 100% and the battery capacity is greater than 99%.

Battery test until battery low can be started anytime.

Tested on Masterguard A1000 and A2000 units.

Signed-off-by: Atanas Vladimirov <vladimirov.atanas@gmail.com>